Propper Baking pipe! @MrTJP
CCQuadBakery is set out like the old tessellator from 1.7, set any
colour / brightness/ normal then addVertex handles actually placing all
in to a holder quad for baking, As soon as a CCQuad is full it is baked
in to a BakedQuad ready for being returned through finishBaking, Finish
baking will throw an exception if , A) a quad is still being baked, or
B) there are no baked quads.
CCQuadBakery can take as many quads as you can throw at it as it will
return a List of BakedQuads!
CCQuads when baked will automatically quadulate and compute normals if
required.