CCL's ASM is now GONE!
CCL is still a coremod as it has ASM libs that need to be there really
early. CCL's IItemRenderer still functions normally but now uses an
overriden RenderItem class to sniff for IItemRenderer models. In
addition i have added support for ANY baked model that is rendererd as
an item to implement IGLTransform and IMatrixTransform, So you can now
apply transforms yourself or give it a CCL Matrix4.
Also, Before IItemRenderer is called to render it will push and pop the
GL state using GlStateManagerHelper to avoid any GL leaks.