Fix for severe World object memory leak
Using objects that do not override hashCode and equals methods correctly as the key of a Map can be a bad idea. The dimension id is a unique identifier of a dimension, so it's appropriate to use this instead. Having the World object on the HashMap prevents the garbage collector to dispose the World object appropriately.