Fix PropertyInteger.parseValue only handling ints
This catches the potential NumberFormatException that might get thrown if some-one passes something that isn't an integer, such as "hello" into the function. This now matches the behaviour of vanilla minecraft's PropertyInteger.parseValue, which catches the error.

I'm not sure when this could happen in vanilla, however it looks like some commands and advancement criteria call this in order to parse their respective strings.

This was found with an in-dev version of buildcraft.