Remove wrong OpenGL version checks for uniforms/shader types Fixes: #412
These are not accurate as the GL caps only report what the context was constructed with. Other GL extensions, etc, may provide support for these various features, or the context may be constructed with forward support enabled. There is no replacement for this, do your own checks before using any of these types.